pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend="hatchling.build"

[project]
name = "iommi-demo"
authors = [{name = "Sam Mangham", email = "mangham@gmail.com"}]
description = "Demo project for the Iommi web framework"
readme = "README.md"
license = { text = "Apache-2.0" }
version = "0.1.0"
requires-python = ">=3.11"
keywords = []

classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "dash-bootstrap-templates>=2.1.0",
    "django>=5.2.1",
    "django-allauth[saml,socialaccount]>=65.12.1",
    "django-auditlog>=3.3.0",
    "django-cors-headers>=4.9.0",
    "importlib>=1.0.4",
    "iommi>=7.13.0",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "plotly>=6.4.0",
    "pyparsing>=3.2.3",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "rules>=3.5",
    "setuptools>=80.9.0",
]

[project.optional-dependencies]
develop = [
    "build",
    "bump-my-version",
    "check-manifest",
    "codespell>=2.4,<2.5",
    "hatchling",
    "mdformat>=0.7.22,<1.1",
    "mdformat-tables>=1",
    "pytest",
    "pytest-cov",
    "ruff>=0.9,<0.15",
    "twine",
    "uv",
    "wheel",
    "django-fastdev",
    "django-pycharm-breakpoint",
    "parso>=0.8.5",
]

[project.scripts]
djmanage = "iommi_demo.manage:main"

[project.urls]
Repository = "https://github.com/smangham/iommi-demo"
Homepage = "https://github.com/smangham/iommi-demo"

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true
commit_args = "-s"

[[tool.bumpversion.files]]
filename = "iommi_demo/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[tool.check-manifest]
ignore = [
    ".copier-answers.yaml",
    "Makefile",
    "docs/**/*",
]

[tool.coverage.run]
branch = true
omit = [
    "iommi_demo/tests/integration/",
]
[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]
ignore_errors = true
fail_under = 50

[tool.hatch.build]
artifacts = []

[tool.hatch.build.sources]
src = "/"

[tool.hatch.build.targets.sdist]
packages = ["iommi_demo"]

[tool.hatch.build.targets.wheel]
packages = ["iommi_demo"]

[tool.hatch.build.targets.wheel.shared-data]

[tool.pytest.ini_options]
addopts = ["-vvv", "--junitxml=junit.xml"]
testpaths = "iommi_demo/tests"

[tool.ruff]
line-length = 150

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
combine-as-imports = true
default-section = "third-party"
known-first-party = ["iommi_demo", "app"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.yardang]
title = "iommi-demo"
root = "README.md"
pages = []
use-autoapi = true
